phase-1-speed-u[

+++ src/server/f5_tts/model/backbones/dit.py
@@ -60,7 +60,7 @@ class TextEmbedding(nn.Module):
             # sinus pos emb
             batch_start = torch.zeros((batch,), dtype=torch.long)
             pos_idx = get_pos_embed_indices(batch_start, seq_len, max_pos=self.precompute_max_pos)
-            text_pos_embed = self.freqs_cis[pos_idx]
+            text_pos_embed = self.freqs_cis[pos_idx].to(text.dtype)
             text = text + text_pos_embed
 
             # convnextv2 blocks
